Move SumoXml and add impl to SumoCfg
