+++ src/main/kotlin/app/urbanflo/urbanflosumoserver/model/network/SumoConnectionsXml.kt
@@ -1,5 +1,6 @@
 package app.urbanflo.urbanflosumoserver.model.network
 
+import app.urbanflo.urbanflosumoserver.model.SumoXml
 import app.urbanflo.urbanflosumoserver.simulation.SimulationId
 import com.fasterxml.jackson.dataformat.xml.annotation.JacksonXmlElementWrapper
 import com.fasterxml.jackson.dataformat.xml.annotation.JacksonXmlProperty
