+++ src/main/kotlin/app/urbanflo/urbanflosumoserver/model/sumocfg/SumoCfg.kt
@@ -1,5 +1,7 @@
 package app.urbanflo.urbanflosumoserver.model.sumocfg
 
+import app.urbanflo.urbanflosumoserver.model.SumoXml
+import app.urbanflo.urbanflosumoserver.simulation.SimulationId
 import com.fasterxml.jackson.dataformat.xml.annotation.JacksonXmlRootElement
 import java.nio.file.Path
 
@@ -7,7 +9,7 @@ import java.nio.file.Path
 class SumoCfg(
     netPath: Path,
     rouPath: Path
-) {
+): SumoXml {
     val input: SumoCfgInput
 
     init {
@@ -16,4 +18,6 @@ class SumoCfg(
             SumoCfgRouteFiles(rouPath.fileName.toString())
         )
     }
+
+    override fun fileName(simulationId: SimulationId) = "$simulationId.sumocfg"
 }
